pytest fuxture moved to conftest.py and pytest-cov added to requirements
