+++ .gitignore
@@ -5,6 +5,9 @@
 # Environment
 env/
 
+# Test coverage
+.coverage
+
 # Images and avatars
 src/images/*
 src/avatars/*
